Automatically generate test data when development
