+++ app.py
@@ -1,4 +1,4 @@
-from api.app import create_app
+from api.app import create_app, initdb, generate
 from cryptography.fernet import Fernet
 
 app = create_app({
@@ -7,50 +7,11 @@ app = create_app({
     'SQLALCHEMY_DATABASE_URI': 'mysql+mysqlconnector://root:password@mysql/db',
 })
 
+@app.cli.command("initdb")
+def initdb_():
+    initdb();
 
-@app.cli.command()
-def initdb():
-    from api.models import db
-    db.create_all()
+@app.cli.command("generate")
+def generate_():
+    generate();
 
-
-@app.cli.command()
-def generate():
-    from api.models import Lottery, Classroom, User, db
-    from werkzeug.security import generate_password_hash
-
-    total_index = 4
-    grades = [5, 6]
-
-    def classloop(f):
-        for grade in grades:
-            for class_index in range(4):  # 0->A 1->B 2->C 3->D
-                f(grade, class_index)
-
-    def create_classrooms(grade, class_index):
-        room = Classroom(grade=grade, index=class_index)
-        db.session.add(room)
-
-    def create_lotteries(grade, class_index):
-        room = Classroom.query.filter_by(
-            grade=grade, index=class_index).first()
-        for perf_index in range(total_index):
-            lottery = Lottery(classroom_id=room.id, index=perf_index, done=False)
-            db.session.add(lottery)
-
-    classloop(create_classrooms)
-    db.session.commit()
-    classloop(create_lotteries)
-    db.session.commit()
-
-    def make_debug_user(name):
-        user = User(username=name, passhash=generate_password_hash(name))
-        db.session.add(user)
-        db.session.commit()
-        return user
-
-    make_debug_user('admin')
-    for i in range(5):
-        make_debug_user(f"example{i}")
-
-    db.session.commit()
